fix vulnerability (.s), And add feature to cancel sending a fkle

+++ index.js
@@ -183,7 +183,7 @@ bot.on("message", (nick, to, text) => {
           [
             "Hello. I'm a file editor bot.",
             "To start editing, type .e [filename]",
-            "You can use .ls, .rn, or .rm to manage your file.",
+            "You can use .ls, .rn, .rm, .i, and .s to manage your files.",
           ].join("\n")
         );
         break;
@@ -241,6 +241,10 @@ bot.on("message", (nick, to, text) => {
             return bot.say(to, "share: Usage: .s [to] [filename]");
           if (!inbox.has(uname)) inbox.set(uname, []);
 
+          if (inbox.get(uname).filter(i =>
+            i.fromUser === nick && i.filename === filename
+          ).length > 0) return bot.say(to, "share: That file is already in queue.");
+
           let n = inbox.get(uname).push({
             fromUser: nick,
             filename: filename,
@@ -249,7 +253,30 @@ bot.on("message", (nick, to, text) => {
           bot.notice(uname, `${nick} is sending you a file: ${filename}`);
           bot.notice(uname, `To accept this, Type .i ${n}`);
           bot.notice(uname, `To refuse this request, Type .ri ${n}`);
-          bot.say(to, "share: " + uname);
+          bot.say(to, `share: Request sent to ${uname}.`);
+          bot.say(to, `share: To cancel sending this file, Send ".cs ${uname} ${filename}"`);
+        }
+        break;
+      case ".cs":
+        {
+          let splitted = name.split(" ");
+          let uname = splitted[0];
+          let filename = splitted.slice(1).join(" ");
+          if (filename.includes("..") || filename.includes("/"))
+            return bot.say(to, "cancel_send: Illegal file name");
+          if (!inbox.has(uname) || !filename) return bot.say(to, "cancel_send: Usage: .cs [username] [filename]");
+
+          let uib = inbox.get(uname).filter(i =>
+            i.fromUser === nick && i.filename === filename
+          );
+
+          if (!uib.length) return bot.say(to, `cancel_send: You didn't send ${filename} to ${uname}`);
+          inbox.set(uname, inbox.get(uname).filter(i => {
+            i.fromUser !== nick && i.filename !== filename
+          }));
+
+          bot.say(to, "cancel_send: Succesfully cancelled");
+          bot.notice(uname, `cancel_send: ${nick} cancelled to send ${filename}.`);
         }
         break;
       case ".i":
@@ -258,22 +285,23 @@ bot.on("message", (nick, to, text) => {
           if (!inbox.has(nick) || !inbox.get(nick).length)
             return bot.say(to, "inbox: empty");
           if (!name) {
+            if (nick !== to) bot.say(to, `${nick}: Inbox list has been sent to PM.`);
             bot.say(
-              to,
+              nick,
               `Inbox: You have ${inbox.get(nick).length} pending requests`
             );
 
             inbox
               .get(nick)
               .forEach((i, index) =>
-                bot.say(to, `${index + 1}| From ${i.fromUser}: ${i.filename}`)
+                bot.say(nick, `${index + 1}| From ${i.fromUser}: ${i.filename}`)
               );
 
             bot.say(
-              to,
+              nick,
               "End of Inbox. To accept one of those code sending request, Type .i [num]"
             );
-            bot.say(to, "To Reject request, Do .ri [num]");
+            bot.say(nick, "To Reject request, Do .ri [num]");
           } else if (parseInt(name, 10) > 0) {
             let i = inbox.get(nick)[parseInt(name, 10) - 1];
             if (!i) return bot.say(to, "Inbox: Empty");
@@ -285,6 +313,14 @@ bot.on("message", (nick, to, text) => {
                   i.fromUser + "_" + i.filename
                 }), Skipped`
               );
+
+            if (!fs.existsSync(__dirname + "/.codes/" + i.fromUser + "/" + i.filename)) {
+              bot.say(to, `share: ${i.fromUser} deleted ${i.filename}. Operation cancelled.`);
+              bot.notice(i.fromUser, `share: You deleted ${i.filename} before ${nick} accept your request. Operation cancelled.`);
+
+              delete inbox.get(nick)[parseInt(name, 10) - 1];
+              return inbox.set(nick, inbox.get(nick).flat());
+            };
             fs.cpSync(
               __dirname + "/.codes/" + i.fromUser + "/" + i.filename,
               userdir + "/" + i.fromUser + "_" + i.filename
@@ -373,4 +409,11 @@ bot.on("message", (nick, to, text) => {
   }
 });
 
+bot.on("registered", () => console.log("login: Succesfully logged as", config.nick));
+bot.on("motd", console.log);
 bot.on("error", console.error);
+
+process.on('SIGINT', () => {
+  console.log("\nWaiting for server to disconnect....");
+  bot.disconnect("Editor Quit");
+});
